move RecrefAdapter to recref_adapter.py
